add tests and fixes project
