+++ README.md
@@ -6,6 +6,7 @@
 * цена (price) тип "float"
 * количество в наличии (quantity) тип "int"
 
+Добавлена функция сложения цен одинаковых категорий
 ### Создан класс Category
 Класс имеет следующие свойства
 * название (name) тип "str"
@@ -32,7 +33,8 @@ poetry install
 
 
 ## Тестирование
-* Добавлены тесты модуля create_class
+* Добавлены тесты модуля categories.py
+* Добавлены тесты модуля product.py
 
-Покрытие тестами составляет 98%
+Покрытие тестами составляет 100%
