+++ htmlcov/index.html
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.2">coverage.py v7.6.2</a>,
-            created at 2024-10-19 23:50 +0700
+            created at 2024-10-24 22:15 +0700
         </p>
     </div>
 </header>
@@ -78,20 +78,27 @@
                 <td class="right" data-ratio="0 0">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_create_class_py.html">src\create_class.py</a></td>
-                <td>47</td>
+                <td class="name left"><a href="z_145eef247bfb46b6_categories_py.html">src\categories.py</a></td>
+                <td>28</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="47 47">100%</td>
+                <td class="right" data-ratio="28 28">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_product_py.html">src\product.py</a></td>
+                <td>29</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="29 29">100%</td>
             </tr>
         </tbody>
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>47</td>
+                <td>57</td>
                 <td>0</td>
                 <td>0</td>
-                <td class="right" data-ratio="47 47">100%</td>
+                <td class="right" data-ratio="57 57">100%</td>
             </tr>
         </tfoot>
     </table>
@@ -103,11 +110,11 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.6.2">coverage.py v7.6.2</a>,
-            created at 2024-10-19 23:50 +0700
+            created at 2024-10-24 22:15 +0700
         </p>
     </div>
     <aside class="hidden">
-        <a id="prevFileLink" class="nav" href="z_145eef247bfb46b6_create_class_py.html"></a>
+        <a id="prevFileLink" class="nav" href="z_145eef247bfb46b6_product_py.html"></a>
         <a id="nextFileLink" class="nav" href="z_145eef247bfb46b6___init___py.html"></a>
         <button type="button" class="button_prev_file" data-shortcut="["></button>
         <button type="button" class="button_next_file" data-shortcut="]"></button>
