+++ tests/conftest.py
@@ -37,3 +37,13 @@ def new_product() -> dict:
         "price": 180000.0,
         "quantity": 5,
     }
+
+
+@pytest.fixture
+def product_1():
+    return Product("Iphone 15", "512GB, Gray space", 210000.0, 8)
+
+
+@pytest.fixture
+def product_2():
+    return Product("Samsung Galaxy C23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
